health, imu, log, and test update with new uint format

+++ tests/unit/utils/mock_helpers.hpp
@@ -17,14 +17,14 @@ extern "C" {
 // helpers for the timer_get_ms function
 
 // w_status_t timer_get_ms(float *time_ms);
-FAKE_VALUE_FUNC(w_status_t, timer_get_ms, float *);
+FAKE_VALUE_FUNC(w_status_t, timer_get_ms, uint32_t *);
 
 // static makes this accessible to only the *_test.cpp file that includes this header.
 // test file should modify this variable to set the timer value as needed
-static float mock_timer_ms = 0.0f;
+static uint32_t mock_timer_ms = 0;
 
 // in test setups, set the custom_fake of timer_get_ms to this function
-static w_status_t timer_get_ms_custom_fake(float *time_ms) {
+static w_status_t timer_get_ms_custom_fake(uint32_t *time_ms) {
     *time_ms = mock_timer_ms;
     return W_SUCCESS;
 }
